improving scaling issues

+++ src/shaders/declarations.wgsl
@@ -54,9 +54,9 @@ struct GaussianBlur {
 
 struct FxIO {
     in_idx: u32,
-    in_downscale: f32,
+    in_downscale: u32,
     out_idx: u32,
-    out_downscale: f32,
+    out_downscale: u32,
 }
 
 fn pi() -> f32 {
